Address feedbacks regarding integration test setup
